refactor: 채팅 날짜 포매팅 함수 분리
